feat: Add ability to extend built-in elements.
This is to support customised build-ins at some point, but to also provide custom base classes to
the main component class.

Fixes #804
